+++ src/cap-project.ts
@@ -1,0 +1,11 @@
+export interface CapacitorProjectState
+{
+	iosBundleId?: string;
+	androidBundleId?: string;
+	iosVersion?: string;
+	androidVersion?: string;
+	iosBuild?: number;
+	androidBuild?: number;
+	iosDisplayName?: string;
+	androidDisplayName?: string;
+}