+++ src/ionic-tree-provider.ts
@@ -10,7 +10,7 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
 	private _onDidChangeTreeData: vscode.EventEmitter<Recommendation | undefined | void> = new vscode.EventEmitter<Recommendation | undefined | void>();
 	readonly onDidChangeTreeData: vscode.Event<Recommendation | undefined | void> = this._onDidChangeTreeData.event;
 
-	constructor(private workspaceRoot: string | undefined, private extensionPath: string) {
+	constructor(private workspaceRoot: string | undefined, private context: vscode.ExtensionContext) {
 	}
 
 	refresh(): void {
@@ -29,10 +29,10 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
 
 		if (element) {
 			return Promise.resolve(element.children);
-		} else {
+		} else {			
 			const packageJsonPath = path.join(this.workspaceRoot, 'package.json');
 			if (this.pathExists(packageJsonPath)) {
-				return reviewProject(this.workspaceRoot, this.extensionPath);
+				return reviewProject(this.workspaceRoot, this.context);
 			} else {
 				return Promise.resolve(starterProject(this.workspaceRoot));
 			}